requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
asttokens==3.0.0
async-timeout==4.0.3
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ipython==8.35.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.3.27
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.28
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.1
tavily-python==0.5.4
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.23.0
